+++ redbot/core/drivers/postgres/postgres.py
@@ -22,7 +22,7 @@ DROP_DDL_SCRIPT_PATH = _PKG_PATH / "drop_ddl.sql"
 
 
 def encode_identifier_data(
-    id_data: IdentifierData
+    id_data: IdentifierData,
 ) -> Tuple[str, str, str, List[str], List[str], int, bool]:
     return (
         id_data.cog_name,
